refactor: 💡 make export/import, add env example
