+++ _env
@@ -1,3 +1,8 @@
 NOTION_KEY=secret_***
 # example https://www.notion.so/hogehoge/{THIS}?v=fugafuga
 NOTION_DATABASE=****
+REDIS_DB=1
+AWS_ID=
+AWS_SECRET=
+AWS_REGION=
+AWS_BUCKET_NAME=kibela-to-notion
